+++ src/app/material-timepicker/components/ngx-material-timepicker-content/ngx-material-timepicker-content.component.ts
@@ -6,5 +6,5 @@ import { Component, Input } from '@angular/core';
 })
 export class NgxMaterialTimepickerContentComponent {
     @Input() appendToInput: boolean;
-    @Input() inputElement: HTMLInputElement;
+    @Input() inputElement: any;
 }